hardware/requirements.txt
flask
RPi.GPIO
pyserial
picamera
base64